cleanup / refactoring
